Group classes into 'duke' package
