+++ data/duke.txt
@@ -1,3 +1,3 @@
-T | 0 | work
-D | 0 | presentation | Feb 03 2024
-E | 0 | party | 4pm | 10pm
+T | 0 | pilates
+T | 0 | swimming
+D | 0 | modelling | May 08 2024
