+++ src/main/java/duke/ToDo.java
@@ -1,4 +1,4 @@
-import java.util.ArrayList;
+package duke;
 
 public class ToDo extends Task {
